+++ .prettierrc.json
@@ -31,6 +31,7 @@
     "",
     "^@/components/ui/(.*)$",
     "^@/components/(.*)$",
+    "^@/components",
     "",
     "^@/styles/(.*)$",
     "",
