+++ src/app/brand/page.tsx
@@ -3,10 +3,7 @@ import { Metadata } from "next";
 
 import cx from "@/utils/cx";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
-import CopyButton from "@/components/copy-button";
-import PageHeaderTitle from "@/components/page-header-title";
+import {Bg, Container, CopyButton, PageHeaderTitle} from "@/components";
 
 export const metadata: Metadata = {
   title: "Brand Assets",
