+++ src/app/careers/[slug]/page.tsx
@@ -4,11 +4,9 @@ import { SITE_URL } from "@/utils/const";
 import { allJobs } from "@content";
 import type { Job } from "@content";
 
-import Bg from "@/components/bg";
 import CareerHeader from "@/components/career/header";
-import Container from "@/components/container";
-import PageBodyGradient from "@/components/page-body-gradient";
-import { Mdx } from "@/components/post/mdx";
+import { Mdx } from "@/components/post";
+import { Bg, Container, PageBodyGradient } from "@/components";
 
 type Props = {
   params: {
