+++ src/app/careers/page.tsx
@@ -3,10 +3,7 @@ import Link from "next/link";
 
 import { allJobs } from "@content";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
-import PageHeaderDesc from "@/components/page-header-desc";
-import PageHeaderTitle from "@/components/page-header-title";
+import {Bg, Container, PageHeaderDesc, PageHeaderTitle} from "@/components";
 
 const description =
   "Help us build the cutting edge data platform for the serverless era.";
