+++ src/app/examples/page.tsx
@@ -1,10 +1,7 @@
 import React from "react";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
 import { Client } from "@/components/example/client";
-import PageHeaderDesc from "@/components/page-header-desc";
-import PageHeaderTitle from "@/components/page-header-title";
+import { Bg, Container, PageHeaderDesc, PageHeaderTitle } from "@/components";
 
 import { getData, type Example } from "./get-data";
