+++ src/app/pricing/layout.tsx
@@ -3,10 +3,7 @@ import "./page.css";
 import { ReactNode } from "react";
 import { Metadata } from "next";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
-import PageHeaderDesc from "@/components/page-header-desc";
-import PageHeaderTitle from "@/components/page-header-title";
+import { Bg, Container, PageHeaderDesc, PageHeaderTitle } from "@/components";
 
 export const metadata: Metadata = {
   title: "Pricing",
