+++ src/app/pricing/qstash/page.tsx
@@ -3,14 +3,14 @@
 import QStashFaqJson from "@/../public/faq/qstash.json";
 import { generateFaqSchema } from "@/utils/structured-schema-generators";
 
-import Container from "@/components/container";
-import PageHeaderDesc from "@/components/page-header-desc";
-import PageHeaderTitle from "@/components/page-header-title";
-import Enterprise from "@/components/pricing/enterprise";
-import ProductToggle from "@/components/pricing/product-toggle";
-import CompareTable from "@/components/pricing/qstash/compare-table";
-import FAQ from "@/components/pricing/qstash/faq";
-import PricingTable from "@/components/pricing/qstash/pricing-table";
+import {
+  CompareTable,
+  Enterprise,
+  FAQ,
+  PricingTable,
+  ProductToggle,
+} from "@/components/pricing";
+import { Container, PageHeaderDesc, PageHeaderTitle } from "@/components";
 
 export default function PricingQStashPage() {
   const structuredFaqSchema = generateFaqSchema(QStashFaqJson);
