+++ src/app/pricing/redis/page.tsx
@@ -3,14 +3,14 @@
 import RedisFaqJson from "@/../public/faq/redis.json";
 import { generateFaqSchema } from "@/utils/structured-schema-generators";
 
-import Container from "@/components/container";
-import PageHeaderDesc from "@/components/page-header-desc";
-import PageHeaderTitle from "@/components/page-header-title";
-import Enterprise from "@/components/pricing/enterprise";
-import ProductToggle from "@/components/pricing/product-toggle";
-import CompareTable from "@/components/pricing/redis/compare-table";
-import FAQ from "@/components/pricing/redis/faq";
-import PricingTable from "@/components/pricing/redis/pricing-table";
+import {
+  CompareTable,
+  Enterprise,
+  FAQ,
+  PricingTable,
+  ProductToggle,
+} from "@/components/pricing";
+import { Container, PageHeaderDesc, PageHeaderTitle } from "@/components";
 
 export default function PricingRedisPage() {
   const structuredFaqSchema = generateFaqSchema(RedisFaqJson);
