+++ src/components/blog/index.ts
@@ -1,0 +1,3 @@
+export { default as PostGridCard } from "./grid-item";
+export { default as PopularTag } from "./popular-tag";
+export { default as PostListCard } from "./list-item";
