+++ src/components/example/comp.tsx
@@ -6,8 +6,7 @@ import Image from "next/image";
 import cx from "@/utils/cx";
 import Balancer from "react-wrap-balancer";
 
-import IconQStash from "@/components/icon-qstash";
-import IconRedis from "@/components/icon-redis";
+import { IconQStash, IconRedis } from "@/components";
 
 export function Example({
   className,
