+++ src/components/example/filter.tsx
@@ -6,8 +6,7 @@ import { HTMLProps } from "react";
 import cx from "@/utils/cx";
 
 import Icon, { ICON_NAMES } from "@/components/icon";
-import IconQStash from "@/components/icon-qstash";
-import IconRedis from "@/components/icon-redis";
+import {IconQStash, IconRedis} from "@/components";
 
 export const ProductsLabel = {
   redis: "Redis",
