+++ src/components/home/community/comp.tsx
@@ -2,7 +2,7 @@ import { HTMLProps } from "react";
 
 import cx from "@/utils/cx";
 
-import Button from "@/components/button";
+import {Button} from "@/components";
 
 export function CommunityBox({
   children,
