+++ src/components/home/investor/index.tsx
@@ -2,14 +2,13 @@ import Link from "next/link";
 
 import { HOME_SECTIONS } from "@/utils/const";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
 import {
   SectionHeader,
   SectionHeaderSummary,
   SectionHeaderTitle,
 } from "@/components/home/section-header";
-import Investors from "@/components/investor/investors";
+import { Investors } from "@/components/investor";
+import { Bg, Container } from "@/components";
 
 export default function HomeInvestors() {
   return (
