+++ src/components/home/price/index.tsx
@@ -1,7 +1,5 @@
 import { HOME_SECTIONS } from "@/utils/const";
 
-import Bg from "@/components/bg";
-import Container from "@/components/container";
 import PriceQStash from "@/components/home/price/qstash";
 import PriceRedis from "@/components/home/price/redis";
 import {
@@ -10,6 +8,7 @@ import {
   SectionHeaderTitle,
 } from "@/components/home/section-header";
 import Icon, { ICON_NAMES } from "@/components/icon";
+import { Bg, Container } from "@/components";
 
 export default function HomeServerless() {
   return (
