+++ src/components/home/product/qstash.tsx
@@ -6,8 +6,7 @@ import { Product } from "@/utils/type";
 
 import { useSegment } from "@/hooks/use-segment";
 
-import Button from "@/components/button";
-import IconQStash from "@/components/icon-qstash";
+import { Button, IconQStash } from "@/components";
 
 import {
   ProductBox,
