+++ src/components/index.ts
@@ -1,0 +1,12 @@
+export { default as Container } from "./container";
+export { default as PageHeaderDesc } from "./page-header-desc";
+export { default as PageHeaderTitle } from "./page-header-title";
+export { default as Analytics } from "./Analytics";
+export { default as ThirdPartyScripts } from "./scripts";
+export { default as Bg } from "./bg";
+export { default as Button } from "./button";
+export { default as IconQStash } from "./icon-qstash";
+export { default as IconRedis } from "./icon-redis";
+export { default as IconVector } from "./icon-vector";
+export { default as PageBodyGradient } from "./page-body-gradient";
+export { default as CopyButton } from "./copy-button";
