+++ src/components/investor/index.ts
@@ -1,0 +1,2 @@
+export { default as Investors } from "./investors";
+export * as InvestorsIcons from "./icons";
