+++ src/components/master/index.ts
@@ -1,0 +1,3 @@
+export {default as Footer} from './footer'
+export {default as Header} from './header'
+export {default as HeaderMobile} from './header-mobile'
