+++ src/components/post/index.ts
@@ -1,0 +1,5 @@
+export { Mdx } from "./mdx";
+export { default as Clap } from "./claps";
+export { default as PostHeader } from "./header";
+export { default as OtherPostCard } from "./other-post";
+export { default as PostTags } from "./tags";
