+++ src/components/pricing/index.ts
@@ -1,0 +1,5 @@
+export { default as ProductToggle } from "./product-toggle";
+export { default as CompareTable } from "./vector/compare-table";
+export { default as Enterprise } from "./vector/enterprise";
+export { default as FAQ } from "./vector/faq";
+export { default as PricingTable } from "./vector/pricing-table";
