+++ src/lib/segment/provider.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { PropsWithChildren, useEffect, useMemo } from "react";
+import React, { PropsWithChildren, useMemo } from "react";
 
 import { Segment } from "./segment";
